Decouple Nexus features into a subpackage (WIP)
