+++ stalk/io/__init__.py
@@ -5,23 +5,15 @@ __email__ = "tiihonen@iki.fi"
 __license__ = "BSD-3-Clause"
 
 from .GeometryLoader import GeometryLoader
-from .NexusGenerator import NexusGenerator
 from .PesLoader import PesLoader
 from .FilesFunction import FilesFunction
 from .FilesLoader import FilesLoader
 from .XyzGeometry import XyzGeometry
-from .PwscfGeometry import PwscfGeometry
-from .PwscfPes import PwscfPes
-from .QmcPes import QmcPes
 
 __all__ = [
     'GeometryLoader',
     'PesLoader',
-    'XyzGeometry',
-    'PwscfGeometry',
-    'NexusGenerator',
     'FilesLoader',
     'FilesFunction',
-    'PwscfPes',
-    'QmcPes'
+    'XyzGeometry',
 ]
