+++ stalk/nexus/NexusGenerator.py
@@ -1,7 +1,10 @@
+from structure import Structure
+
+
 class NexusGenerator():
     '''A wrapper class for generating Nexus functions to produce and represent a PES.'''
     func = None
-    args = None
+    args = {}
 
     def __init__(self, func, args={}):
         '''A Nexus PES function is constructed from the job-generating function and arguments.'''
@@ -11,12 +14,12 @@ class NexusGenerator():
 
     def generate(self, structure, path, **kwargs):
         '''Return a list of Nexus jobs provided "structure" and "path" arguments.'''
-        # Make a copy of the structure in case it is changed in the function.
-        structure_job = structure.copy()
-        structure_job.to_nexus_only()
+        if not isinstance(structure, Structure):
+            raise TypeError('Must use Structure class with NexusGenerator')
+        # end if
         args = self.args.copy()
         args.update(kwargs)
-        jobs = self.func(structure_job, path, **args)
+        jobs = self.func(structure, path, **args)
         return jobs
     # end def
