+++ stalk/nexus/PwscfGeometry.py
@@ -1,11 +1,12 @@
-from .GeometryLoader import GeometryLoader
+from nexus import PwscfAnalyzer
+
+from ..io.GeometryLoader import GeometryLoader
 from stalk.params.GeometryResult import GeometryResult
 
 
 class PwscfGeometry(GeometryLoader):
 
-    def __load__(self, path, suffix='relax.in', c_pos=1.0, **kwargs):
-        from nexus import PwscfAnalyzer
+    def __load__(self, path, suffix='relax.in', c_pos=1.0, **kwargs):        
         ai = PwscfAnalyzer('{}/{}'.format(path, suffix), **kwargs)
         ai.analyze()
         pos = ai.structures[len(ai.structures) - 1].positions * c_pos
