+++ stalk/nexus/PwscfPes.py
@@ -1,11 +1,13 @@
+
+from nexus import PwscfAnalyzer
+
 from stalk.params.PesResult import PesResult
-from .PesLoader import PesLoader
+from ..io.PesLoader import PesLoader
 
 
 class PwscfPes(PesLoader):
 
     def __load__(self, path, suffix='scf.in', **kwargs):
-        from nexus import PwscfAnalyzer
         ai = PwscfAnalyzer('{}/{}'.format(path, suffix), **kwargs)
         ai.analyze()
         E = ai.E
