+++ stalk/nexus/QmcPes.py
@@ -1,11 +1,13 @@
+
+from nexus import QmcpackAnalyzer
+
 from stalk.params.PesResult import PesResult
-from .PesLoader import PesLoader
+from ..io.PesLoader import PesLoader
 
 
 class QmcPes(PesLoader):
 
     def __load__(self, path, qmc_idx=1, suffix='dmc/dmc.in.xml', **kwargs):
-        from nexus import QmcpackAnalyzer
         ai = QmcpackAnalyzer('{}/{}'.format(path, suffix), **kwargs)
         ai.analyze()
         LE = ai.qmc[qmc_idx].scalars.LocalEnergy
